Fix maximize/restore glitch of OpaqueBrowserFrameView in Weston.
Make the opaque view layout to be updated once window states are
updated upon the host window manager events like if a user moves a window
in such a way so that it becomes automatically maximized or restored.

The layout update behaviour is copied from DesktopWindowTreeHostX11::OnWmStateChanged.
Now DesktopWindowTreeHostMus subscribes for a root window, which receives
window state updates, in addition to the content window, and invalidates
and updates client and non-client views once aura::client::kShowStateKey
key is received.

Issue #270