Make ozone backends to propage activation state to aura
This patch makes it possible to propagate activation state from
ozone backends to aura by add additional method OnActivationChanged
to WindowTree/WindowTreeClient.

The patch works the following way:
There are two cases:
1) There is only one browser window.
2) There are two or more browser windows. Basically, this doesn't
differ that much from the first case.

In first case, when the browser window looses focus (client clicks
outside the browser area or clicks other (browser) window) or gains focus
(client clicks on the browser window. It can be non-client or client area),
ozone backends call their delegate that they lost or gained focus,
which results in a call to WindowTreeClient::OnActivationChanged.

The WindowTreeClient calls aura::FocusSynchronizer with or without
active focus client, which sets or resets an activation client resulting
in a call to FocusSynchronizer::OnActiveFocusClientChanged and
FocusSynchronizerObserver::OnActiveFocusClientChanged. The first one
instantiates a FlightChange for focus and calls to WindowTree::SetFocus,
the second one results in DesktopWindowTreeHostX11::OnActiveFocusClientChanged,
which results in a call to DesktopNativeWidgetAura::HandleActivationChanged.

The last one updates the views then and other things related to lost or
gained focus.

fixup! Make ozone backends to propage activation state to aura

Handle pointer grab properly and add missing implementation. I messed up a
bit and broke it.

What's broken:

When user opens a browser window and presses inside it, they are not able to
click anywhere else after that as long as pointer grab is not unset.

fixup! Make ozone backends to propage activation state to aura

Don't put window mapped or focused on initialization as long as
those will be set on configure event. This fixes focusing in X11, when
a new window is created.