[DoNoCarryForward] [ozone/x11] events can be dispatched to the wrong window
cherry-picked from: https://crrev.com/c/774778

When multiple X11Window/X11WindowOzone instances are created, each registers
itself in PlatformEventSource as a "platform event dispatcher".

Before dispatching an event to a specific window, PlatformEventSource
iterates over the list of registered dispatchers, "asking" each dispatcher
whether the event can be dispatch to it.

In case of non-Ozone X11 windows, X11Window::CanDispatchEvent is called
receiving a XEvent* parameter, and checks whether the xwindow targeted by the
event is the same as the xwindow it wraps (see::IsEventForXWindow).  If so it
dispatches the event to this X11Window instance.

In case of Ozone X11 windows, X11WindowOzone::CanDispatchEvent is called
receiving a ui::Event* parameter, and it is not possible to access its
associated NativeEvent*, because event was manually constructed in
EventSourceLibevent::ProcessXEvent.  Hence, ::CanDispatchEvent does not check if
it is the event's target window, but instead whether the event location is
contained in its bounds.

This is fragile: if there are two overlapping X11WindowOzone instances, the
first one that registered itself in PlatformEventSource will be the one
receiving mouse events in the overlapping area, even if it is not the foreground
window.

CL fixes it by extending XEventDispatcher interface with two new methods:
::CheckCanDispatchNextPlatformEvent and ::PlatformEventDispatchFinished.

Once the XEvent is translated to an ui::Event, XEventDispatchers are
iterated and asked to ::CheckCanDispatchNextPlatformEvent if they can process
the translated event based on XEvent::XID. If so, they must make an internal
promise to process next translated event sent by PlatformEventSource.  Once
done, ::PlatformEventDispatchFinished is called and the promise must be reset.

Originally based on https://codereview.chromium.org/2791693003/

BUG=707406

Change-Id: I2bfd1f3f0f5ed574e7ad9634ade7020d5a99dba5