fixup! Handle events for interactive drag and resize of native windows.
When starting an interactive drag or resize of native windows,
the first thing called is WaylandPointer::Button, and depending on
what mouse button is pressed, we set the appropriate flag to
WaylandPointer::flag_.

The problem is that when aura/mus informs WS that it is going to enter
an interactive move/resize, and this information gets to Ozone, the
"mouse up" event of the original "mouse down" isn't set.

This means than, when entering interactive move/resize mode, we left
WaylandPointer::flags_ set.

Now say user ends the interactive move/resize, and starts to hover around with
mouse into NON CLIENT areas, for resizing the window. It happens that mouse cursor
does not change to "resize types" when getting close the edge of the browser window.
This is because WaylandPointer::Motion is sending "mouse move" events
with WaylandPointer::flags_ still set. On Aura side, these events are
converted to "dragging" events (see ui/events/event.cc snippet below).

  case ET_POINTER_MOVED:
    if (pointer_event.flags() &
        (EF_LEFT_MOUSE_BUTTON | EF_MIDDLE_MOUSE_BUTTON |
         EF_RIGHT_MOUSE_BUTTON)) {
      SetType(ET_MOUSE_DRAGGED);
    } else {
      SetType(ET_MOUSE_MOVED);
    }
    break;

... and in CompoundEventFilter::OnMouseEvent, ::UpdateCursor is not called.

Patch fixes this by resetting WaylandPointer::flags_ whenever entering
interactive move/resize mode.

Issue #314