fixup! c++ / mojo changes for 'external window mode'
Patch moves our external window entry point from
WTC::CreateWindowPortForTopLevel to WTC::OnWindowMusCreated.

The later is the same spot used in ChromeOS/Mus to create
the WindowManager root window, so we concentrate changes and
logic in one place.

The early return in ::CreateWindowPortForTopLevel, is because
we do not want to set server_id there, otherwise the first line in
::OnWindowMusCreated bails out.

Issue #266