[ozone/wayland] Fix KeyboardEvdev: stop autorepating and fix modifiers
A couple of points were missed, when we switched to use KeyboardEvdev.

First of all, when user switches between windows by pressing alt+tab,
KeyboardEvdev may find itself in a situation, when auto repeating of
the alt key is scheduled over and over again. In order to fix this,
KeyboardEvdev::StopRepeating is called on Leave event.

Secondly, once user alternates between windows and then returns back
to the same window, a text input may be broken due to set modidifers.
That is, when user switches to another window, the alt key is still
treated pressed and modifiers are not updated at all before user
switches back and presses the alt key again. The fix to this problem
is to reset modifiers on Leave event.

Last but not least, WaylandKeyboard::RepeatInfo sends delay and rate
information that is usefull for KeyboardEvdev. Thus, desired repeating
settings can be set. What is more, wayland documentation says if
rate is send as 0, then auto repeat should be turned off.

Issue #331