[ozone/wayland] use KeyboardEvDev for keyboard handling
In WaylandKeyboard::Key' body, we deliberately duplicate some of the
logic in KeyboardEvdev::DispatchKey. The idea was to use the bits of
KeyboardEvDev needed for proper keyboard input handling.
It turns out that features like key repeat and specific cases involving
modifiers were not functional.

Instead of continue to duplicate stable and existing logic, CL adopts
the use of KeyboardEvDev from WaylandKeyboard, fixing all these issues
and being able to do some code clean ups (follow ups).

CL also changes the KeyboardLayoutManager instantiation order in
OzonePlatformWayland::InitializeUI, so that it happens prior to
WaylandConnection instantiation.
Reason: once WaylandConnection creates WaylandKeyboard and (implicitly)
KeyboardEvDev, the later needs the KeyboardLayoutEngine properly set.

Issue #331

fixup! [ozone/wayland] use KeyboardEvDev for keyboard handling

A couple of points were missed, when we switched to use KeyboardEvdev.

First of all, when user switches between windows by pressing alt+tab,
KeyboardEvdev may find itself in a situation, when auto repeating of
the alt key is scheduled over and over again. In order to fix this,
KeyboardEvdev::StopRepeating is called on Leave event.

Secondly, once user alternates between windows and then returns back
to the same window, a text input may be broken due to set modidifers.
That is, when user switches to another window, the alt key is still
treated pressed and modifiers are not updated at all before user
switches back and presses the alt key again. The fix to this problem
is to reset modifiers on Leave event.

Last but not least, WaylandKeyboard::RepeatInfo sends delay and rate
information that is usefull for KeyboardEvdev. Thus, desired repeating
settings can be set. What is more, wayland documentation says if
rate is send as 0, then auto repeat should be turned off.