Add a mock impl XKBWaylandKeyboardLayoutEngine
In order to allow existing unittests (namely WaylandKeyboardTests)
and future ones to pass with XKB enabled, we need a specialization of
XkbWaylandKeyboardLayoutEngine for tests.
Following the logic used throughout the Wayland unittests' machinery,
it makes sense to mock a Xkb-capable KeyboardLayoutEngine, rather than
using an actual implementation.

This CL adds a mock implementation of XkbWaylandKeyboardLayoutEngine to
be used by Wayland unittests.

For this, it turns XkbWaylandKeyboardLayoutEngine into a pure virtual
class, which is implemented by XkbWaylandKeyboardLayoutEngineImpl
(for regular ozone/wayland execution, e.g. chrome/wayland), and
adds a sibling MockXkbWaylandKeyboardLayoutEngine derived class to
be used for unittests.

BUG=578890

Change-Id: Ie98f0d5bcf5b45c86812c4fab7dc63e1ffdc0890