Implement filter to run content browser tests with Linux/mus build.
This filter disables all the crashing and failing tests and
leaves only the working ones. Some of the tests are broken
both with stock Chromium build and Chromium/mus build.

fixup! Implement filter to run content browser tests with Linux/mus build.

Moving the filter file to a right location.